Moved code from Vector2Tester in src source folder to Vector2Test in test source folder.
